+++ src/id_translation/dio/integration/dask.py
@@ -3,23 +3,56 @@
 import typing as _t
 from collections import abc as _abc
 
+import pandas as _pd
 from dask import compute as _compute
 from dask import dataframe as _dd
+from rics.misc import tname as _tname
 
-from id_translation import dio as _dio
 from id_translation import types as _tt
+from id_translation.dio import DataStructureIO as _DataStructureIO
 from id_translation.dio.exceptions import NotInplaceTranslatableError as _NotInplaceTranslatableError
-from id_translation.offline import MagicDict as _MagicDict
+from id_translation.dio.integration.pandas import PandasIO as _PandasIO
 from id_translation.offline import TranslationMap as _TranslationMap
 
 DaskT = _t.TypeVar("DaskT", _dd.DataFrame, _dd.Series)
 """Supported ``dask`` types."""
 
+PartitionT = _t.TypeVar("PartitionT", _pd.DataFrame, _pd.Series)
+"""Supported ``dask`` partition types."""
+PartitionIO = _PandasIO[PartitionT, str, _tt.SourceType, _tt.IdType]
+"""A ``dask`` partition IO implementation."""
 
-class DaskIO(_dio.DataStructureIO[DaskT, str, _tt.SourceType, _tt.IdType]):
-    """Optional IO implementation for ``dask`` types."""
+del _PandasIO  # Don't use by accident.
+
+
+class DaskIO(_DataStructureIO[DaskT, str, _tt.SourceType, _tt.IdType]):
+    """Optional IO implementation for ``dask`` types.
+
+    Args:
+        missing_as_nan: If set, unknown IDs will be `NaN`. Forwarded to :class:`.PandasIO`.
+        as_category: Set `dtype='category'` in the result. Forwarded to :class:`.PandasIO`.
+
+    Notes:
+        Combining ``missing_as_nan=False`` with ``as_category=True`` can be unpredictable in distributed contexts.
+    """
 
     priority = 1980
+
+    def __init__(
+        self,
+        *,
+        missing_as_nan: bool | None = None,
+        as_category: bool = False,
+    ) -> None:
+        self._part_io = PartitionIO[_t.Any, _tt.SourceType, _tt.IdType](
+            missing_as_nan=missing_as_nan,
+            as_category=as_category,
+        )
+
+    @property
+    def partition_io(self) -> PartitionIO[_t.Any, _tt.SourceType, _tt.IdType]:
+        """The :class:`PartitionIO` implementation used by this instance."""
+        return self._part_io
 
     @classmethod
     def handles_type(cls, arg: _t.Any) -> bool:
@@ -36,49 +69,41 @@ class DaskIO(_dio.DataStructureIO[DaskT, str, _tt.SourceType, _tt.IdType]):
     def extract(cls, translatable: DaskT, names: list[str]) -> dict[str, _abc.Sequence[_tt.IdType]]:
         if isinstance(translatable, _dd.Series):
             if len(names) != 1:
-                raise RuntimeError(f"{len(names)=} != 1 is not supported for dask.Series")
+                msg = f"{len(names)=} != 1 is not supported for {_tname(translatable, include_module=True)}"
+                raise NotImplementedError(msg)
             return {names[0]: translatable.unique().compute().to_list()}
         else:
             delayed = {n: translatable[n].unique() for n in names}
             name_to_ids = _compute(delayed)[0]
             return {name: ids.to_list() for name, ids in name_to_ids.items()}
 
-    @classmethod
     def insert(
-        cls,
+        self,
         translatable: DaskT,
         names: list[str],
         tmap: _TranslationMap[str, _tt.SourceType, _tt.IdType],
         copy: bool,
-    ) -> DaskT | None:
+    ) -> DaskT:
         if not copy:
             raise _NotInplaceTranslatableError(translatable)  # Can't in-place a compute graph.
 
-        if isinstance(translatable, _dd.Series):
-            return _translate_series(translatable, tmap[names[0]])
-        else:
-            return _translate_frame(translatable, names, tmap)
+        return translatable.map_partitions(  # type: ignore[no-any-return]
+            translate_partition,
+            names,
+            tmap,
+            self.partition_io,
+        )
 
 
-def _translate_series(series: _dd.Series, magic_dict: _MagicDict[_tt.IdType]) -> _dd.Series:
-    mapping = {idx: magic_dict[idx] for idx in series.unique().compute()}
-    return series.replace(mapping)  # type: ignore[no-any-return]
-
-
-def _translate_frame(
-    df: _dd.DataFrame,
-    names: list[_tt.NameType],
-    tmap: _TranslationMap[_tt.NameType, _tt.SourceType, _tt.IdType],
-) -> _dd.DataFrame:
-    original_columns = df.columns
-
-    try:
-        df.columns = [str(i) for i in range(len(original_columns))]
-        for name, int_col in zip(original_columns, df.columns, strict=True):
-            if name in names:
-                translated = _translate_series(df[int_col], tmap[name])
-                df[int_col] = translated
-    finally:
-        df.columns = original_columns
-
-    return df
+def translate_partition(
+    part: PartitionT,
+    names: list[str],
+    tmap: _TranslationMap[str, _tt.SourceType, _tt.IdType],
+    part_io: PartitionIO[PartitionT, _tt.SourceType, _tt.IdType],
+) -> PartitionT:
+    """Translation a single Dask partition."""
+    if isinstance(part, _pd.DataFrame):
+        part_io.insert(part, names, tmap, copy=False)
+        return part
+    else:
+        return part_io.insert(part, names, tmap, copy=True)